+++ CHANGELOG.rst
@@ -13,6 +13,10 @@ Added
 Changed
 ^^^^^^^
 * `@senwu`_: Change learning framework from Tensorflow to PyTorch.
+* `@lukehsiao`_: Rather than importing ambiguously from ``fonduer`` directly,
+  disperse imports into their respective pipeline phases. This eliminates
+  circular dependencies, and makes imports more explicit and clearer to the
+  user where each import is originating from.
 * `@lukehsiao`_: Provide debug logging of external subprocess calls.
 * `@lukehsiao`_: Use ``tdqm`` for progress bar (including multiprocessing).
 * `@lukehsiao`_: Set the default PostgreSQL client encoding to "utf8".
@@ -90,6 +94,59 @@ Fixed
     Furthermore, because Candidates are now composed of Mentions rather than
     directly of Spans, to get the Span object from a mention, use the ``.span``
     attribute of a Mention.
+
+.. note::
+    Fonduer has been reorganized to require more explicit import syntax. In
+    Fonduer v0.2.3, nearly everything was imported directly from fonduer:
+
+    .. code:: python
+
+        from fonduer import (
+            CandidateExtractor,
+            DictionaryMatch,
+            Document,
+            FeatureAnnotator,
+            GenerativeModel,
+            HTMLDocPreprocessor,
+            Intersect,
+            LabelAnnotator,
+            LambdaFunctionMatcher,
+            MentionExtractor,
+            Meta,
+            Parser,
+            RegexMatchSpan,
+            Sentence,
+            SparseLogisticRegression,
+            Union,
+            candidate_subclass,
+            load_gold_labels,
+            mention_subclass,
+        )
+
+    With this release, you will now import from each pipeline phase. This makes
+    imports more explicit and allows you to more clearly see which pipeline
+    phase each import is associated with:
+
+    .. code:: python
+
+        from fonduer import Meta
+        from fonduer.candidates import (
+            CandidateExtractor,
+            DictionaryMatch,
+            Intersect,
+            LambdaFunctionMatcher,
+            MentionExtractor,
+            RegexMatchSpan,
+            Union,
+        )
+        from fonduer.candidates.models import candidate_subclass, mention_subclass
+        from fonduer.features import FeatureAnnotator
+        from fonduer.learning import GenerativeModel, SparseLogisticRegression
+        from fonduer.parser import Parser
+        from fonduer.parser.models import Document, Sentence
+        from fonduer.parser.preprocessors import HTMLDocPreprocessor
+        from fonduer.supervision import LabelAnnotator, load_gold_labels
+
 
 [0.2.3] - 2018-07-23
 --------------------
