+++ fonduer/learning/utils.py
@@ -9,11 +9,80 @@ import numpy as np
 import scipy.sparse as sparse
 from pandas import DataFrame
 
+from fonduer.learning.models import Marginal
+from fonduer.utils.utils_annotations import csr_AnnotationMatrix
+
 logger = logging.getLogger(__name__)
 
 # ###########################################################
 # # General Learning Utilities
 # ###########################################################
+
+
+def save_marginals(session, X, marginals, training=True):
+    """Save marginal probabilities for a set of Candidates to db.
+
+    :param X: Either an M x N csr_AnnotationMatrix-class matrix, where M
+        is number of candidates, N number of LFs/features; OR a list of
+        arbitrary objects with candidate ids accessible via a .id attrib
+    :param marginals: A dense M x K matrix of marginal probabilities, where
+        K is the cardinality of the candidates, OR a M-dim list/array if K=2.
+    :param training: If True, these are training marginals / labels; else they
+        are saved as end model predictions.
+
+    Note: The marginals for k=0 are not stored, only for k = 1,...,K
+    """
+    logger = logging.getLogger(__name__)
+    # Make sure that we are working with a numpy array
+    try:
+        shape = marginals.shape
+    except Exception as e:
+        marginals = np.array(marginals)
+        shape = marginals.shape
+
+    # Handle binary input as M x 1-dim array; assume elements represent
+    # poksitive (k=1) class values
+    if len(shape) == 1:
+        marginals = np.vstack([1 - marginals, marginals]).T
+
+    # Only add values for classes k=1,...,K
+    marginal_tuples = []
+    for i in range(shape[0]):
+        for k in range(1, shape[1] if len(shape) > 1 else 2):
+            if marginals[i, k] > 0:
+                marginal_tuples.append((i, k, marginals[i, k]))
+
+    # NOTE: This will delete all existing marginals of type `training`
+    session.query(Marginal).filter(Marginal.training == training).delete(
+        synchronize_session="fetch"
+    )
+
+    # Prepare bulk INSERT query
+    q = Marginal.__table__.insert()
+
+    # Check whether X is an AnnotationMatrix or not
+    anno_matrix = isinstance(X, csr_AnnotationMatrix)
+    if not anno_matrix:
+        X = list(X)
+
+    # Prepare values
+    insert_vals = []
+    for i, k, p in marginal_tuples:
+        cid = X.get_candidate(session, i).id if anno_matrix else X[i].id
+        insert_vals.append(
+            {
+                "candidate_id": cid,
+                "training": training,
+                "value": k,
+                # We cast p in case its a numpy type, which psycopg2 does not handle
+                "probability": float(p),
+            }
+        )
+
+    # Execute update
+    session.execute(q, insert_vals)
+    session.commit()
+    logger.info("Saved {%d} marginals".format(len(marginals)))
 
 
 def reshape_marginals(marginals):
