--- fonduer/utils/__init__.py
@@ -1,27 +1,0 @@
-from fonduer.utils.utils import (
-    camel_to_under,
-    get_as_dict,
-    matrix_conflicts,
-    matrix_coverage,
-    matrix_fn,
-    matrix_fp,
-    matrix_overlaps,
-    matrix_tn,
-    matrix_tp,
-    remove_files,
-    tokens_to_ngrams,
-)
-
-__all__ = [
-    "camel_to_under",
-    "get_as_dict",
-    "matrix_conflicts",
-    "matrix_coverage",
-    "matrix_fn",
-    "matrix_fp",
-    "matrix_overlaps",
-    "matrix_tn",
-    "matrix_tp",
-    "remove_files",
-    "tokens_to_ngrams",
-]
