+++ fonduer/utils/data_model_utils/visual.py
@@ -6,8 +6,8 @@ from builtins import range
 from collections import defaultdict
 
 from fonduer.candidates import Ngrams
-from fonduer.utils import tokens_to_ngrams
 from fonduer.utils.data_model_utils.utils import _to_span, _to_spans
+from fonduer.utils.utils import tokens_to_ngrams
 from fonduer.utils.utils_visual import (
     bbox_from_sentence,
     bbox_from_span,
